requirements.txt
torch==2.1.0
transformers==4.36.0
accelerate==0.25.0
bitsandbytes==0.41.3
peft==0.7.0
datasets==2.15.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
chromadb==0.4.18
rank-bm25==0.2.2
pypdf==3.17.0
python-docx==1.1.0
beautifulsoup4==4.12.2
lxml==4.9.3
streamlit==1.29.0
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.6
python-dotenv==1.0.0
tqdm==4.66.1
pandas==2.1.4
numpy==1.24.3